almost implement global transaction screen

+++ app/src/main/java/com/ithirteeng/secondpatternsclientproject/features/client/myloans/main/presentation/model/MyLoansMainEffect.kt
@@ -4,9 +4,9 @@ import com.ithirteeng.secondpatternsclientproject.common.architecture.BaseEffect
 
 sealed interface MyLoansMainEffect : BaseEffect {
 
-    data class NavigateToProgramInfoScreen(val programId: Int) : MyLoansMainEffect
+    data class NavigateToProgramInfoScreen(val programId: Long) : MyLoansMainEffect
 
-    data class NavigateToLoanInfoScreen(val loanId: Int) : MyLoansMainEffect
+    data class NavigateToLoanInfoScreen(val loanId: Long) : MyLoansMainEffect
 
     data class ShowError(val message: String) : MyLoansMainEffect
 }
